+++ noodles-sam/CHANGELOG.md
@@ -10,6 +10,8 @@
   * sam/record: Add `Record::parse_with_fields` to decode only a subset of
     fields.
 
+  * sam/record: Add template length wrapper (`TemplateLength`).
+
 ### Changed
 
   * sam/record/cigar: Change conversion from `Vec<Op>` to be